requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
bcrypt>=4.0.0
PyJWT>=2.0.0
pydantic[email]
